Delete all the things
