+++ compiler/server/Server/Endpoints/Project/CreateProject.hs
@@ -9,13 +9,13 @@ module Server.Endpoints.Project.CreateProject
     CreateProject,
   )
 where
+import Server.Helpers.ProjectData
 
 import qualified Data.Aeson as JSON
 import Data.OpenApi
 import GHC.Generics
 import Language.Mimsa.Project.Stdlib
 import Servant
-import Server.Handlers
 import Server.Types
 
 type CreateProject =
