+++ compiler/server/Server/Endpoints/Project/Evaluate.hs
@@ -17,13 +17,9 @@ import qualified Data.Aeson as JSON
 import Data.OpenApi hiding (Server, get)
 import Data.Text (Text)
 import GHC.Generics
-import qualified Language.Mimsa.Actions.Evaluate as Actions
-import qualified Language.Mimsa.Actions.Helpers.CheckStoreExpression as Actions
 import qualified Language.Mimsa.Actions.Helpers.Parse as Actions
 import qualified Language.Mimsa.Actions.Modules.Evaluate as Actions
-import qualified Language.Mimsa.Actions.Monad as Actions
 import Language.Mimsa.Printer
-import Language.Mimsa.Transform.Warnings
 import Language.Mimsa.Types.Project
 import Language.Mimsa.Types.Store
 import Servant
@@ -34,54 +30,12 @@ import Server.Types
 
 -- using https://github.com/haskell-servant/servant/blob/master/doc/cookbook/uverb/UVerb.lhs
 
-type EvaluateAPI = EvaluateSE :<|> EvaluateModule
+type EvaluateAPI = EvaluateModule
 
 evaluate :: MimsaEnvironment -> Server EvaluateAPI
-evaluate mimsaEnv =
-  evaluateExpression mimsaEnv :<|> evaluateModuleExpression mimsaEnv
+evaluate =
+  evaluateModuleExpression
 
--- /project/evaluate/
-
-type EvaluateSE =
-  "evaluate"
-    :> ReqBody '[JSON] EvaluateRequest
-    :> JsonPost EvaluateResponse
-
-data EvaluateRequest = EvaluateRequest
-  { erCode :: Text,
-    erProjectHash :: ProjectHash
-  }
-  deriving stock (Eq, Ord, Show, Generic)
-  deriving anyclass (JSON.FromJSON, ToSchema)
-
-data EvaluateResponse = EvaluateResponse
-  { erResult :: Text,
-    erExpressionData :: ExpressionData
-  }
-  deriving stock (Eq, Ord, Show, Generic)
-  deriving anyclass (JSON.ToJSON, ToSchema)
-
-evaluateExpression ::
-  MimsaEnvironment ->
-  EvaluateRequest ->
-  MimsaHandler EvaluateResponse
-evaluateExpression mimsaEnv (EvaluateRequest code hash) =
-  runMimsaHandlerT $ do
-    let action = do
-          expr <- Actions.parseExpr code
-          (_, simpleExpr, se, typedExpr, input) <-
-            Actions.evaluate code expr
-          project <- Actions.getProject
-          res <- Actions.checkStoreExpression input project se
-          let warnings = getWarnings res
-          pure $
-            EvaluateResponse
-              (prettyPrint simpleExpr)
-              (makeExpressionData se typedExpr input warnings)
-    response <- lift $ eitherFromActionM mimsaEnv hash action
-    case response of
-      Left e -> throwMimsaError e
-      Right (_, _, a) -> returnMimsa a
 
 -- /project/evaluate-module/
