+++ compiler/server/Server/Servant.hs
@@ -11,7 +11,6 @@ where
 import Data.Proxy
 import Servant
 import Server.Endpoints.Compile
-import Server.Endpoints.Expression
 import Server.Endpoints.Module
 import Server.Endpoints.Project
 import Server.Types
@@ -19,7 +18,6 @@ import Server.Types
 type MimsaAPI =
   ProjectAPI
     :<|> CompileAPI
-    :<|> ExpressionAPI
     :<|> ModuleAPI
 
 mimsaAPI :: Proxy MimsaAPI
@@ -29,5 +27,4 @@ mimsaServer :: MimsaEnvironment -> Server MimsaAPI
 mimsaServer mimsaEnv =
   projectEndpoints mimsaEnv
     :<|> compileEndpoints mimsaEnv
-    :<|> expressionEndpoints mimsaEnv
     :<|> moduleEndpoints mimsaEnv
