--- compiler/src/Language/Mimsa/Actions/BindExpression.hs
@@ -1,56 +1,0 @@
-{-# LANGUAGE OverloadedStrings #-}
-
-module Language.Mimsa.Actions.BindExpression (bindExpression) where
-
-import Data.Foldable (traverse_)
-import Data.Text (Text)
-import qualified Language.Mimsa.Actions.Helpers.CheckStoreExpression as Actions
-import qualified Language.Mimsa.Actions.Helpers.FindExistingBinding as Actions
-import qualified Language.Mimsa.Actions.Monad as Actions
-import qualified Language.Mimsa.Actions.Typecheck as Actions
-import Language.Mimsa.Printer
-import Language.Mimsa.Project.Helpers
-import Language.Mimsa.Store
-import Language.Mimsa.Transform.Warnings
-import Language.Mimsa.Types.AST
-import Language.Mimsa.Types.Identifiers
-import Language.Mimsa.Types.ResolvedExpression
-import Language.Mimsa.Types.Store
-
--- bind a new expression to the project
-bindExpression ::
-  Expr Name Annotation ->
-  Name ->
-  Text ->
-  Actions.ActionM (ExprHash, ResolvedExpression Annotation)
-bindExpression expr name input = do
-  project <- Actions.getProject
-  -- if there is an existing binding of this name, bring its deps into scope
-  -- perhaps this should be more specific and actually take the previous
-  -- exprHash as an argument instead of the name
-  resolvedExpr <- case Actions.findExistingBinding name project of
-    -- there is an existing one, use its deps when evaluating
-    Just se ->
-      let newSe = setStoreExpression se expr
-       in Actions.checkStoreExpression input project newSe
-    -- no existing binding, resolve as usual
-    Nothing ->
-      Actions.typecheckExpression project input expr
-  let storeExpr = reStoreExpression resolvedExpr
-
-  -- print any warnings
-  traverse_ (Actions.appendMessage . prettyPrint) (getWarnings resolvedExpr)
-
-  Actions.bindStoreExpression storeExpr name
-  case lookupBindingName project name of
-    Nothing -> do
-      Actions.appendMessage
-        ( "Bound " <> prettyPrint name <> "."
-        )
-      pure (getStoreExpressionHash storeExpr, resolvedExpr)
-    Just _ ->
-      do
-        Actions.appendMessage
-          ( "Updated binding of " <> prettyPrint name <> "."
-          )
-        pure (getStoreExpressionHash storeExpr, resolvedExpr)
