--- compiler/src/Language/Mimsa/Actions/Evaluate.hs
@@ -1,88 +1,0 @@
-{-# LANGUAGE OverloadedStrings #-}
-
-module Language.Mimsa.Actions.Evaluate
-  ( evaluate,
-  )
-where
-
-import Control.Monad.Except
-import Data.Bifunctor
-import Data.Foldable (traverse_)
-import qualified Data.Map.Strict as M
-import Data.Text (Text)
-import qualified Language.Mimsa.Actions.Helpers.CheckStoreExpression as Actions
-import qualified Language.Mimsa.Actions.Helpers.GetDepsForStoreExpression as Actions
-import qualified Language.Mimsa.Actions.Interpret as Actions
-import qualified Language.Mimsa.Actions.Monad as Actions
-import qualified Language.Mimsa.Actions.Optimise as Actions
-import qualified Language.Mimsa.Actions.Typecheck as Actions
-import Language.Mimsa.Printer
-import Language.Mimsa.Store
-import Language.Mimsa.Transform.Warnings
-import Language.Mimsa.Types.AST
-import Language.Mimsa.Types.Error
-import Language.Mimsa.Types.Identifiers
-import Language.Mimsa.Types.ResolvedExpression
-import Language.Mimsa.Types.Store
-import Language.Mimsa.Types.Typechecker
-import Prettyprinter
-
-evaluate ::
-  Text ->
-  Expr Name Annotation ->
-  Actions.ActionM
-    ( MonoType,
-      Expr Name Annotation,
-      StoreExpression Annotation,
-      Expr Name MonoType,
-      Text
-    )
-evaluate input expr = do
-  project <- Actions.getProject
-  -- typecheck expression
-  resolved <-
-    Actions.typecheckExpression project input expr
-
-  let se = reStoreExpression resolved
-
-  -- get dependencies of StoreExpression
-  depsSe <- Actions.getDepsForStoreExpression se
-
-  -- optimise them all like a big legend
-  storeExprs <- Actions.optimiseAll (fst <$> depsSe)
-
-  -- get new root StoreExpression (it may be different due to optimisation)
-  optimisedStoreExpr <- case M.lookup (getStoreExpressionHash se) storeExprs of
-    Just re -> pure re
-    _ -> throwError (StoreErr (CouldNotFindStoreExpression (getStoreExpressionHash se)))
-
-  -- resolve optimised expression
-  (ResolvedExpression mt newStoreExpr _expr' typedExpr input') <-
-    Actions.checkStoreExpression
-      (prettyPrint optimisedStoreExpr)
-      project
-      optimisedStoreExpr
-
-  -- expr with types and Name
-  let typedNameExpr = first fst typedExpr
-
-  -- interpret
-  interpretedExpr <-
-    Actions.interpreter newStoreExpr
-
-  -- print any warnings
-  traverse_ (Actions.appendMessage . prettyPrint) (getWarnings resolved)
-
-  -- print
-  Actions.appendDocMessage
-    ( group
-        ( prettyDoc interpretedExpr
-            <> line
-            <> "::"
-            <> line
-            <> prettyDoc mt
-        )
-    )
-  pure (mt, interpretedExpr, newStoreExpr, typedNameExpr, input')
-
----------
