+++ compiler/src/Language/Mimsa/Actions/Helpers/LookupExpression.hs
@@ -1,13 +1,12 @@
 module Language.Mimsa.Actions.Helpers.LookupExpression
-  ( lookupExpressionInStore,
-    lookupExpression,
+  (
+
     lookupModule,
     lookupModuleByName,
   )
 where
 
 import Control.Monad.Except
-import Data.Functor
 import qualified Data.Map.Strict as M
 import qualified Language.Mimsa.Actions.Monad as Actions
 import Language.Mimsa.Project.Helpers
@@ -15,7 +14,6 @@ import Language.Mimsa.Types.AST
 import Language.Mimsa.Types.Error
 import Language.Mimsa.Types.Modules
 import Language.Mimsa.Types.Project
-import Language.Mimsa.Types.Store
 
 lookupModuleByName :: ModuleName -> Actions.ActionM (Module Annotation)
 lookupModuleByName modName = do
@@ -32,21 +30,3 @@ lookupModule modHash = do
   case M.lookup modHash (prjModuleStore project) of
     Just mod' -> pure mod'
     Nothing -> throwError (StoreErr (CouldNotFindModule modHash))
-
-lookupExpression ::
-  ExprHash ->
-  Actions.ActionM (StoreExpression Annotation)
-lookupExpression exprHash = do
-  project <- Actions.getProject
-  se <- lookupExpressionInStore (prjStore project) exprHash
-  pure (se $> mempty)
-
--- | given a store, try and find something in it
-lookupExpressionInStore ::
-  Store ann ->
-  ExprHash ->
-  Actions.ActionM (StoreExpression ann)
-lookupExpressionInStore store exprHash =
-  case M.lookup exprHash (getStore store) of
-    Just se -> pure se
-    _ -> throwError (StoreErr (CouldNotFindStoreExpression exprHash))
