+++ compiler/src/Language/Mimsa/Store.hs
@@ -1,6 +1,6 @@
 module Language.Mimsa.Store
   ( module Language.Mimsa.Store.Storage,
-    module Language.Mimsa.Store.Resolver,
+
     module Language.Mimsa.Store.Helpers,
     module Language.Mimsa.Store.ResolvedDeps,
   )
@@ -8,5 +8,4 @@ where
 
 import Language.Mimsa.Store.Helpers
 import Language.Mimsa.Store.ResolvedDeps
-import Language.Mimsa.Store.Resolver
 import Language.Mimsa.Store.Storage
