--- compiler/src/Language/Mimsa/Store/Resolver.hs
@@ -1,87 +1,0 @@
-module Language.Mimsa.Store.Resolver
-  ( extractVars,
-    createStoreExpression,
-  )
-where
-
-import Data.Coerce
-import Data.Map.Strict (Map)
-import qualified Data.Map.Strict as M
-import Data.Maybe (catMaybes, isJust)
-import qualified Data.Set as S
-import Language.Mimsa.Store.ExtractTypes (extractTypes)
-import Language.Mimsa.Store.ExtractVars (extractVars)
-import Language.Mimsa.Typechecker.DataTypes
-import Language.Mimsa.Types.AST
-import Language.Mimsa.Types.Error
-import Language.Mimsa.Types.Identifiers
-import Language.Mimsa.Types.Modules.ModuleName
-import Language.Mimsa.Types.Store
-
--- this takes the expression, works out what it needs from it's environment
--- and wraps that up
--- this would be a good place for a simplifying step in future
--- by replacing all variables internally with a1, a2 etc, we'll get less
--- repetition, ie \x -> x and \y -> y will be the same function and thus hash
-
--- basically, (global Bindings, Expression) -> StoreExpression
-
-findHashInBindings :: Bindings -> Name -> Either ResolverError ExprHash
-findHashInBindings (Bindings bindings') name =
-  case M.lookup name bindings' of
-    Just a -> Right a
-    _ -> Left $ MissingBinding name (Bindings bindings')
-
--- given an expression, and the current environment, create a
--- store expression that captures the hashes of the functions we'll need
-createStoreExpression ::
-  (Eq ann, Monoid ann) =>
-  Bindings ->
-  TypeBindings ->
-  Expr Name ann ->
-  Either ResolverError (StoreExpression ann)
-createStoreExpression bindings' tBindings expr =
-  StoreExpression expr
-    <$> findBindings bindings' expr
-    <*> findTypeBindings tBindings expr
-    <*> pure mempty
-    <*> pure mempty
-
-findBindings ::
-  (Eq ann, Monoid ann) =>
-  Bindings ->
-  Expr Name ann ->
-  Either ResolverError (Map (Maybe ModuleName, Name) ExprHash)
-findBindings bindings' expr = do
-  let findValueHash name =
-        (,) (Nothing, name)
-          <$> findHashInBindings bindings' name
-  valueHashes <- traverse findValueHash (S.toList . extractVars $ expr)
-  pure (M.fromList valueHashes)
-
------------
-
-isBuiltIn :: TyCon -> Bool
-isBuiltIn = isJust . lookupBuiltIn . coerce
-
-findHashInTypeBindings ::
-  TypeBindings ->
-  TyCon ->
-  Either ResolverError (Maybe ExprHash)
-findHashInTypeBindings (TypeBindings bindings') cName =
-  if isBuiltIn cName
-    then Right Nothing
-    else case M.lookup cName bindings' of
-      Just a -> Right (Just a)
-      _ -> Left $ MissingType cName (TypeBindings bindings')
-
-findTypeBindings ::
-  TypeBindings ->
-  Expr Name ann ->
-  Either ResolverError (Map (Maybe ModuleName, TyCon) ExprHash)
-findTypeBindings tBindings expr = do
-  let findTypeHash cName = do
-        maybeHash <- findHashInTypeBindings tBindings cName
-        pure $ (,) (Nothing, cName) <$> maybeHash
-  hashes <- traverse findTypeHash (S.toList . extractTypes $ expr)
-  pure (M.fromList (catMaybes hashes))
